Move ace/pace tests to match new tc module
